lLinting
Former-commit-id: b427eeaf3d4f9d62b239296615ceb2279d786bd4 [formerly 772c617578a23770d952d71f1e542790e1fb2a83]
Former-commit-id: 059bf9f072ae62658498e81a3d67bd55e3b3520c

+++ api/models/categories/delete_test.go
@@ -17,7 +17,6 @@ const (
 )
 
 func (t *CategoriesTestSuite) TestStore_Delete() {
-
 	tt := map[string]struct {
 		want interface{}
 		mock func(m sqlmock.Sqlmock)
@@ -61,7 +60,7 @@ func (t *CategoriesTestSuite) TestStore_Delete() {
 	for name, test := range tt {
 		t.Run(name, func() {
 			s := t.Setup(test.mock)
-			err := s.Delete(int64(category.Id))
+			err := s.Delete(category.Id)
 			if err != nil {
 				t.Contains(errors.Message(err), test.want)
 				return
